requirements.txt
absl-py==2.1.0
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.1.24
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
narwhals==1.25.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
zipp==3.21.0
